pyproject.toml
[project]
name = "objectagentsdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "openai>=1.108.1",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "ultralytics>=8.3.202",
    "bleak>=1.1.1",
    "lap>=0.5.12",
]

[project.optional-dependencies]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    "build",
    "dist",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable essential rules for code quality
select = [
    "E",       # pycodestyle errors
    "W",       # pycodestyle warnings
    "F",       # Pyflakes
    "I",       # isort
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "UP",      # pyupgrade
    "RUF",     # Ruff-specific rules
]

ignore = [
    "E501",    # Line too long (handled by formatter)
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]

[tool.ruff.lint.isort]
known-first-party = ["objectagentsdemo"]
